+++ src/main/java/generaloss/freetype/freetype/FTGlyphMetrics.java
@@ -11,66 +11,66 @@ public class FTGlyphMetrics extends FTStruct { // struct done.
 
 
     // FT_Pos width;
-    private static native long getWidth(long pointer);
+    private static native int getWidth(long pointer);
 
     public float getWidth() {
-        final long raw = getWidth(super.pointer);
+        final int raw = getWidth(super.pointer);
         return FTPos.toFloat(raw);
     }
 
     // FT_Pos height;
-    private static native long getHeight(long pointer);
+    private static native int getHeight(long pointer);
 
     public float getHeight() {
-        final long raw = getHeight(super.pointer);
+        final int raw = getHeight(super.pointer);
         return FTPos.toFloat(raw);
     }
 
     // FT_Pos horiBearingX;
-    private static native long getHoriBearingX(long pointer);
+    private static native int getHoriBearingX(long pointer);
 
     public float getHoriBearingX() {
-        final long raw = getHoriBearingX(super.pointer);
+        final int raw = getHoriBearingX(super.pointer);
         return FTPos.toFloat(raw);
     }
 
     // FT_Pos horiBearingY;
-    private static native long getHoriBearingY(long pointer);
+    private static native int getHoriBearingY(long pointer);
 
     public float getHoriBearingY() {
-        final long raw = getHoriBearingY(super.pointer);
+        final int raw = getHoriBearingY(super.pointer);
         return FTPos.toFloat(raw);
     }
 
     // FT_Pos horiAdvance;
-    private static native long getHoriAdvance(long pointer);
+    private static native int getHoriAdvance(long pointer);
 
     public float getHoriAdvance() {
-        final long raw = getHoriAdvance(super.pointer);
+        final int raw = getHoriAdvance(super.pointer);
         return FTPos.toFloat(raw);
     }
 
     // FT_Pos vertBearingX;
-    private static native long getVertBearingX(long pointer);
+    private static native int getVertBearingX(long pointer);
 
     public float getVertBearingX() {
-        final long raw = getVertBearingX(super.pointer);
+        final int raw = getVertBearingX(super.pointer);
         return FTPos.toFloat(raw);
     }
 
     // FT_Pos vertBearingY;
-    private static native long getVertBearingY(long pointer);
+    private static native int getVertBearingY(long pointer);
 
     public float getVertBearingY() {
-        final long raw = getVertBearingY(super.pointer);
+        final int raw = getVertBearingY(super.pointer);
         return FTPos.toFloat(raw);
     }
 
     // FT_Pos vertAdvance;
-    private static native long getVertAdvance(long pointer);
+    private static native int getVertAdvance(long pointer);
 
     public float getVertAdvance() {
-        final long raw = getVertAdvance(super.pointer);
+        final int raw = getVertAdvance(super.pointer);
         return FTPos.toFloat(raw);
     }
