+++ src/main/java/generaloss/freetype/freetype/FTSizeMetrics.java
@@ -12,64 +12,64 @@ public class FTSizeMetrics extends FTStruct { // struct done.
 
 
     // FT_UShort x_ppem;
-    private static native long getXppem(long pointer);
+    private static native int getXppem(long pointer);
 
-    public long getXppem() {
+    public int getXppem() {
         return getXppem(super.pointer);
     }
 
     // FT_UShort y_ppem;
-    private static native long getYppem(long pointer);
+    private static native int getYppem(long pointer);
 
-    public long getYppem() {
+    public int getYppem() {
         return getYppem(super.pointer);
     }
 
     // FT_Fixed x_scale;
-    private static native long getXScale(long pointer);
+    private static native int getXScale(long pointer);
 
     public float getXScale() {
-        final long raw = getXScale(super.pointer);
+        final int raw = getXScale(super.pointer);
         return FTFixed.toFloat(raw);
     }
 
     // FT_Fixed y_scale;
-    private static native long getYScale(long pointer);
+    private static native int getYScale(long pointer);
 
     public float getYScale() {
-        final long raw = getYScale(super.pointer);
+        final int raw = getYScale(super.pointer);
         return FTFixed.toFloat(raw);
     }
 
     // FT_Pos ascender;
-    private static native long getAscender(long pointer);
+    private static native int getAscender(long pointer);
 
     public float getAscender() {
-        final long raw = getAscender(super.pointer);
+        final int raw = getAscender(super.pointer);
         return FTPos.toFloat(raw);
     }
 
     // FT_Pos descender;
-    private static native long getDescender(long pointer);
+    private static native int getDescender(long pointer);
 
     public float getDescender() {
-        final long raw = getDescender(super.pointer);
+        final int raw = getDescender(super.pointer);
         return FTPos.toFloat(raw);
     }
 
     // FT_Pos height;
-    private static native long getHeight(long pointer);
+    private static native int getHeight(long pointer);
 
     public float getHeight() {
-        final long raw = getHeight(super.pointer);
+        final int raw = getHeight(super.pointer);
         return FTPos.toFloat(raw);
     }
 
     // FT_Pos max_advance;
-    private static native long getMaxAdvance(long pointer);
+    private static native int getMaxAdvance(long pointer);
 
     public float getMaxAdvance() {
-        final long raw = getMaxAdvance(super.pointer);
+        final int raw = getMaxAdvance(super.pointer);
         return FTPos.toFloat(raw);
     }
